Change return of div to float

+++ src/main/java/com/example/springdocker/model/MyMathCalc.java
@@ -10,7 +10,7 @@ public class MyMathCalc {
         return a * b;
     }
 
-    public int divide(int a, int b){
+    public float divide(int a, int b){
         return a / b;
     }
 }
